qing-agent/requirements.txt
# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 数据处理
pydantic>=2.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# AI/ML（可选）
# openai>=1.10.0
# sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0

# 浏览器自动化（可选）
# playwright>=1.40.0

# 工具
python-dotenv>=1.0.0
tqdm>=4.66.0
python-docx>=1.1.0
PyPDF2>=3.0.0

# 缓存（替代 Redis）
cachetools>=5.3.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
requests>=2.31.0
